+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## In-development
 - Fix the default blend mode: it should mix colors using their alphas
+- [Breaking] Remove support for image types other than jpeg and png
 
 ## 0.4.0-alpha0
 The API change is *very breaking*. It can be considered nearly a full re-write of Quicksilver.